+++ MyWebApi/MyWebApi/Interfaces/IUserRepository.cs
@@ -68,12 +68,12 @@ namespace WebApi.Interfaces
         Task<bool> SetUserRtLanguagePrefs(long userId, bool shouldBeConcidered);
         Task<bool> CheckUsersAreCombinableRT(long user1, long user2);
         Task<Balance> GetUserWalletBalance(long userId);
-        Task<int> TopUpPointBalance(long userId, int points, string description);
+        Task<float> TopUpPointBalance(long userId, float points, string description);
         Task<int> TopUpOPBalance(long userId, int points, string description);
         Task<bool> CheckUserHasPremiumAsync(long userId);
         Task<bool> CheckBalanceIsSufficient(long userId, int cost);
         Task<DateTime> GetPremiumExpirationDate(long userId);
-        Task<DateTime> GrantPremiumToUser(long userId, int cost, int dayDuration, Currency currency);
+        Task<DateTime> GrantPremiumToUser(long userId, float cost, int dayDuration, Currency currency);
         Task RegisterUserEncounter(RegisterEncounter model);
         Task<Encounter> GetUserEncounter(long encounterId);
         Task<List<Encounter>> GetUserEncounters(long userId, Section section);
@@ -127,7 +127,7 @@ namespace WebApi.Interfaces
         Task<List<ActiveEffect>> GetUserActiveEffects(long userId);
         Task<bool> DeactivateEffectAsync(long userId, long activeEffectId);
         Task<bool> CheckEffectIsActiveAsync(long userId, Currency effectId);
-        Task<bool> PurchaseEffectAsync(long userId, int effectId, int points, Currency currency, short count=1);
+        Task<bool> PurchaseEffectAsync(long userId, int effectId, float points, Currency currency, short count=1);
         Task<bool> PurchasePersonalityPointsAsync(long userId, int points, Currency currency, short count=1);
         Task PurchasePointsAsync(long userId, int cost, Currency currency, int amount);
         Task<bool> SendTickRequestAsync(SendTickRequest request);
